make parent_ref_targets_server generic
